+++ src/pegr/gensys/PodChunk.cpp
@@ -1,4 +1,4 @@
-#include "pegr/gensys/PodChunkPtr.hpp"
+#include "pegr/gensys/PodChunk.hpp"
 
 namespace pegr {
 namespace Gensys {
